requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
black==22.10.0==py310hbe9552e_0
blis==0.7.8==pypi_0
bzip2==1.0.8==h3422bc3_4
ca-certificates==2022.9.24==h4653dfc_0
catalogue==2.0.8==pypi_0
certifi==2022.9.24==pypi_0
charset-normalizer==2.1.1==pypi_0
click==8.1.3==py310hbe9552e_0
confection==0.0.3==pypi_0
cymem==2.0.6==pypi_0
ffmpeg-python==0.2.0==pypi_0
filelock==3.8.0==pypi_0
future==0.18.2==pypi_0
huggingface-hub==0.10.1==pypi_0
idna==3.4==pypi_0
jinja2==3.1.2==pypi_0
langcodes==3.3.0==pypi_0
libffi==3.4.2==h3422bc3_5
libsqlite==3.39.4==h76d750c_0
libzlib==1.2.12==h03a7124_4
markupsafe==2.1.1==pypi_0
more-itertools==8.14.0==pypi_0
murmurhash==1.0.8==pypi_0
mypy_extensions==0.4.3==py310hbe9552e_5
nb-core-news-sm==3.4.0==pypi_0
ncurses==6.3==h07bb92c_1
numpy==1.23.4==pypi_0
openssl==3.0.5==h03a7124_2
packaging==21.3==pypi_0
pathspec==0.10.1==pyhd8ed1ab_0
pathy==0.6.2==pypi_0
pip==22.2.2==pyhd8ed1ab_0
platformdirs==2.5.2==pyhd8ed1ab_1
preshed==3.0.7==pypi_0
pydantic==1.9.2==pypi_0
pyparsing==3.0.9==pypi_0
python==3.10.6==hae75cb6_0_cpython
python_abi==3.10==2_cp310
pyyaml==6.0==pypi_0
readline==8.1.2==h46ed386_0
regex==2022.9.13==pypi_0
requests==2.28.1==pypi_0
setuptools==65.4.1==pyhd8ed1ab_0
smart-open==5.2.1==pypi_0
spacy==3.4.1==pypi_0
spacy-legacy==3.0.10==pypi_0
spacy-loggers==1.0.3==pypi_0
srsly==2.4.4==pypi_0
thinc==8.1.4==pypi_0
tk==8.6.12==he1e0b03_0
tokenizers==0.13.1==pypi_0
tomli==2.0.1==pyhd8ed1ab_0
torch==1.12.1==pypi_0
tqdm==4.64.1==pypi_0
transformers==4.23.1==pypi_0
typer==0.4.2==pypi_0
typing-extensions==4.4.0==pypi_0
tzdata==2022e==h191b570_0
urllib3==1.26.12==pypi_0
wasabi==0.10.1==pypi_0
wheel==0.37.1==pyhd8ed1ab_0
whisper==1.0==pypi_0
xz==5.2.6==h57fd34a_0
